// type/types.d.ts
export interface carddetails {
  email: string;
  cardNumber: string;
  expiry: string;
  cvc: string;
  billingName: string;
  country: string;
  zipCode: string;
  city: string;
  addressLine1: string;
}

export type threeDotsOptions = 'Copy Interview link' | 'Create talent list' | 'Configure Job Page' | 'Delete opportunity';